Fix issue with build failing on TypeScript error
Bumps the TypeScript version to 3.8.3 (note that this is not the latest
version as of this commit). 3.8.3 doesn't have the spread operator bug.

The build was failing with this message:

```
button-text-card/src/button-text-card.ts(53,7): semantic error TS2783:
'title' is specified more than once, so this usage will be overwritten.
```

Added vscode exclusions to gitignore.